requirements.txt
Python 3.13.2 (tags/v3.13.2:4f8bb39, Feb  4 2025, 15:23:48) [MSC v.1942 64 bit (AMD64)] on win32

numpy 2.1.3
pylink-square 1.6.0
pyserial 3.5
matplotlib 3.10.1

SEGGER J-Link installation (V788 was used at the time)